% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{as_synth}
\alias{as_synth}
\title{as_synth}
\usage{
as_synth(.data, ...)
}
\arguments{
\item{.data}{tibble data frame.}

\item{...}{catch for additional arguments}
}
\value{
An object of class \code{synth_tbl}
}
\description{
Coerce tibble data frame object to class `synth_tbl``
}
