% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_z.R
\name{convert_quo_value}
\alias{convert_quo_value}
\title{Convert Quo Value to Component parts}
\usage{
convert_quo_value(x, quo_list)
}
\arguments{
\item{x}{The value attribute of a quosure produced by the quos function.}
}
\value{
A list with three attributes:  name (character), formula (quosure),
and filter (numeric or date vector)
}
\description{
Each quo value submitted to create_z_matrix has three component parts:  a
variable name, the command, and restictions on the time span.
}
