% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_y.R
\name{create_y_matrix}
\alias{create_y_matrix}
\title{Create Y Matrix}
\usage{
create_y_matrix(data, i_var, t_var, g_def, time_span, y)
}
\arguments{
\item{data}{A data.frame}

\item{i_var}{A character variable with the name of the unit variable (i)}

\item{t_var}{A character variable with the name of the time variable (t)}

\item{g_def}{A character variable with the boolean logic to identify observations
in the treatment group.}

\item{time_span}{An optional vector with up to three values:  The earliest value
of t_var allowed, the latest value that is part of the pre-treatment period,
and the latest time period allowd.}

\item{y}{A character variable with the name of any y variables to include}
}
\description{
This function creats a data.frame that contains the information on the time-
varying variables.
}
