% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_z.R
\name{create_z_matrix}
\alias{create_z_matrix}
\title{Create Z Matrix}
\usage{
create_z_matrix(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
This function creates a Z matrix for the synthetic panel.  It contains one
observation for each unit (i) and one column for each variable to be used as
a control.
}
