% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{generate_predictors}
\alias{generate_predictors}
\title{generate_predictor}
\usage{
generate_predictors(data, time_window = NULL, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
generate_predictor
}
