% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{generate_weights}
\alias{generate_weights}
\title{generate_weights}
\usage{
generate_weights(
  data,
  optimization_window = NULL,
  custom_variable_weights = NULL,
  include_fit = FALSE,
  optimization_method = c("Nelder-Mead", "BFGS"),
  genoud = FALSE,
  quadopt = "ipop",
  Margin.ipop = 5e-04,
  Sigf.ipop = 5,
  Bound.ipop = 10,
  verbose = FALSE,
  track_progress = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
generate_weights
}
