% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{grab_predictor_weights}
\alias{grab_predictor_weights}
\title{grab_predictor_weights}
\usage{
grab_predictor_weights(data, placebo = F)
}
\arguments{
\item{placebo}{}
}
\value{

}
\description{
grab_predictor_weights
}
