% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab-methods.R
\name{grab_signficance}
\alias{grab_signficance}
\title{grab_signficance}
\usage{
grab_signficance(data, time_window = NULL)
}
\arguments{
\item{data}{nested data of type \code{synth_tbl}}

\item{time_window}{time window that the significance values should be
computed.}
}
\value{
tibble data frame contiaing the following fields:
\itemize{
\item \code{unit_name}: name of the unit
\item \code{type}: treated or donor unit (placebo)
\item \code{pre_mspe}: pre-intervention period means squared precise error
\item \code{post_mspe}: post-intervention period means squared precise error
\item \code{mspe_ratio}: post_mspe/pre_mspe; captures the difference in fit in the
pre and post period. A good fit in the pre-period and a poor fit in the
post-period reflects a meaningful effect when comparing the difference
between the observed outcome and the synthetic control.
\item \code{rank}: rank order of the mspe_ratio.
\item \code{fishers_exact_pvalue}: rank/total to generate a p-value. Conventional
levels aren't achievable if there isn't a sufficient number of controls to
generate a large enough ranking. Need at least 20 control units to use the
conventional .05 level.
\item \code{z_score}: (mspe_ratio-mean(mspe_ratio))/sd(mspe_ratio); captures the
degree to which the mspe_ratio of the treated unit deviates from the mean
of the placebo units. Provinding an alternative significance determination.
}
}
\description{
Generate interential statistics comparing the treated unit to the placebo
units.
}
\examples{

\dontrun{

# Smoking example data
data(smoking)

smoking_out <-
smoking \%>\%

# initial the synthetic control object
synthetic_control(outcome = cigsale,
                  unit = state,
                  time = year,
                  i_unit = "California",
                  i_time = 1988,
                  generate_placebos=F) \%>\%

# Generate the aggregate predictors used to generate the weights
  generate_predictor(time_window=1980:1988,
                     lnincome = mean(lnincome, na.rm = T),
                     retprice = mean(retprice, na.rm = T),
                     age15to24 = mean(age15to24, na.rm = T)) \%>\%

  generate_predictor(time_window=1984:1988,
                     beer = mean(beer, na.rm = T)) \%>\%

  generate_predictor(time_window=1975,
                     cigsale_1975 = cigsale) \%>\%

  generate_predictor(time_window=1980,
                     cigsale_1980 = cigsale) \%>\%

  generate_predictor(time_window=1988,
                     cigsale_1988 = cigsale) \%>\%


  # Generate the fitted weights for the synthetic control
  generate_weights(optimization_window =1970:1988,
                   Margin.ipop=.02,Sigf.ipop=7,Bound.ipop=6) \%>\%

  # Generate the synthetic control
  generate_control()

# Plot the observed and synthetic trend
smoking_out \%>\% grab_signficance(time_window = 1970:2000)

}

}
