% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{grab_synthetic_control}
\alias{grab_synthetic_control}
\title{grab_synthetic_control}
\usage{
grab_synthetic_control(data, placebos = T)
}
\arguments{
\item{placebos}{}
}
\value{

}
\description{
grab_synthetic_control
}
