% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{is_synth}
\alias{is_synth}
\title{is_synth}
\usage{
is_synth(x)
}
\arguments{
\item{x}{object to be coerced or tested.}
}
\value{
boolean flag; TRUE if object is of class synth_tbl
}
\description{
test for objects of type "synth_tbl".
}
\examples{

\dontrun{
# Smoking example data
data(smoking)

# initial the synthetic control object
smoking_out <-
smoking \%>\%
synthetic_control(outcome = cigsale,
                  unit = state,
                  time = year,
                  i_unit = "California",
                  i_time = 1988,
                  generate_placebos=F)

# Check if is class synth_tbl
is_synth(smoking_out)
}

}
