% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_z.R
\name{make_z_variable}
\alias{make_z_variable}
\title{Make Individual Z Variable}
\usage{
make_z_variable(quo_list, data, i_name, t_name)
}
\description{
This function takes the inpute produced by convert_quo_value for a single
control variable and returns a data.frame with its value.
}
