% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare Data}
\usage{
prepare_data(data, i, t, treated_def, preperiod, y, ...)
}
\description{
This function processes the data before creating a synthetic panel.
}
