% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_y.R
\name{remove_duplicates}
\alias{remove_duplicates}
\title{Remove Observations (i) with Duplicate Time Values}
\usage{
remove_duplicates(data, i_name, t_name)
}
\description{
This function takes a data.frame and determines whether any of the individuals
}
