% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{remove_incomplete_dates}
\alias{remove_incomplete_dates}
\title{Remove Incomplete Dates}
\usage{
remove_incomplete_dates(data, i_name, t_name, date_vec)
}
\description{
This function removes observations from a data frame that do not have
observations for all of the available dates.
}
