% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth_data.R
\name{synth_data}
\alias{synth_data}
\title{Tidy Data for Synthetic Control}
\usage{
synth_data(formula, data, i_var, t_var, treatment)
}
\arguments{
\item{data}{A data.frame}

\item{i_var}{Name of the variable denoting individual subjects}

\item{t_var}{Name of the variable reflecting time}

\item{treatment}{An NSE statement designating which i are among the treatment (not control) observations}

\item{y}{Names of variable(s)}

\item{z}{A list of control variables.  Each element of the list should be a
vector with 2-3 elements. The first element is the definition of the variable
using NSE.  The second element will be the filter argument and the third
(optional) will be the name for the element.}
}
\description{
Tidy Data for Synthetic Control
}
