% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth.R
\name{synth_method}
\alias{synth_method}
\title{implement_synth}
\usage{
synth_method(
  treatment_unit_covariates = NULL,
  control_units_covariates = NULL,
  control_units_outcome = NULL,
  treatment_unit_outcome = NULL,
  custom.v = NULL,
  optimxmethod = c("Nelder-Mead", "BFGS"),
  genoud = FALSE,
  quadopt = "ipop",
  Margin.ipop = 5e-04,
  Sigf.ipop = 5,
  Bound.ipop = 10,
  verbose = F,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
\link{AUX Function} Original synthetic control method proposed by (Abadie et al.
2003, 2010, 2015) and implemented in \code{synth} package.
}
\details{
Synth works as the main engine of the \code{tidysynth} package. More on the method
and estimation procedures can be found in (Abadie et al. 2010).
}
