% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{synth_weights}
\alias{synth_weights}
\title{synth_weights}
\usage{
synth_weights(
  data,
  time_window = NULL,
  custom_variable_weights = NULL,
  include_fit = FALSE,
  optimization_method = c("Nelder-Mead", "BFGS"),
  genoud = FALSE,
  quadopt = "ipop",
  Margin.ipop = 5e-04,
  Sigf.ipop = 5,
  Bound.ipop = 10,
  verbose = F,
  ...
)
}
\arguments{
\item{custom_variable_weights}{Set custom weights for the covariates}

\item{include_fit}{Return the synth fit in the nested output}

\item{optimization_method}{}
}
\value{

}
\description{
Auxiliary Function for generating individual weights for each data entry.
}
