% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{synthetic_control}
\alias{synthetic_control}
\title{synthetic_control}
\usage{
synthetic_control(
  data = NULL,
  outcome = NULL,
  unit = NULL,
  time = NULL,
  i_unit = NULL,
  i_time = NULL
)
}
\arguments{
\item{i_time}{}
}
\value{

}
\description{
synthetic_control
}
