\name{tidysynth-package}
\alias{tidysynth-package}
\alias{tidysynth}
\docType{package}
\title{
\packageTitle{tidysynth}
}
\description{
\packageDescription{tidysynth}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{tidysynth}
\packageIndices{tidysynth}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{tidysynth}

Maintainer: \packageMaintainer{tidysynth}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the ~~
~~ R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
